requirements.txt
altair==5.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.54.1
frozendict==2.4.5
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mplfinance==0.12.10b0
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.9.1
nbformat==5.10.4
networkx==3.3
numpy==2.1.2
packaging==24.1
pandas==2.2.3
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
protobuf==5.28.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
soupsieve==2.6
statsmodels==0.14.4
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.44
